third_party/blink/web_tests/http/tests/devtools/tracing/timeline-misc/timeline-event-causes-expected.txt

Test that causes are correctly generated for various types of events.


Running: testTimerInstall
PASS - record contained Timer Installed
(anonymous) @ setTimeoutFunction.js:

Running: testRequestAnimationFrame
PASS - record contained Animation Frame Requested
(anonymous) @ requestAnimationFrameFunction.js:

Running: testStyleRecalc
PASS - record contained First Invalidated
styleRecalcFunction @ styleRecalcFunction.js:

Running: testLayout
PASS - record contained Layout Forced
layoutFunction @ layoutFunction.js:
PASS - record contained First Layout Invalidation
layoutFunction @ layoutFunction.js:
